refactor css for layout anime
